Add tests (and reorganize tests a bit)
