+++ python/jupyterlab-collaborative-chat/ui-tests/tests/commands.spec.ts
@@ -1,0 +1,130 @@
+import { expect, IJupyterLabPageFixture, test} from '@jupyterlab/galata';
+
+const FILENAME = 'my-chat.chat';
+
+const fillModal = async (
+  page: IJupyterLabPageFixture,
+  text = '',
+  button: 'Ok' | 'Cancel' = 'Ok'
+): Promise<void> => {
+  const dialog = page.getByRole('dialog');
+  await dialog.getByRole('textbox').pressSequentially(text);
+  await dialog.getByRole('button').filter({ hasText: button }).click();
+};
+
+test.describe('#commandPalette', () => {
+  const name = FILENAME.replace('.chat', '');
+
+  test.beforeEach(async ({ page }) => {
+    await page.keyboard.press('Control+Shift+c');
+  });
+
+  test.afterEach(async ({ page }) => {
+    for (let filename of ['untitled.chat', FILENAME]) {
+      if (await page.filebrowser.contents.fileExists(filename)) {
+        await page.filebrowser.contents.deleteFile(filename);
+      }
+    }
+  });
+
+  test('should have 2 commands in palette', async ({ page }) => {
+    await expect(
+      page.locator(
+        '#modal-command-palette li[data-command^="collaborative-chat"]'
+      )
+    ).toHaveCount(2);
+  });
+
+  test('should create a chat with name from command palette', async ({
+    page
+  }) => {
+    await page
+      .locator(
+        '#modal-command-palette li[data-command="collaborative-chat:create"]'
+      )
+      .click();
+    await fillModal(page, name);
+    await page.waitForCondition(
+      async () => await page.filebrowser.contents.fileExists(FILENAME)
+    );
+    await expect(page.activity.getTabLocator(FILENAME)).toBeVisible();
+  });
+
+  test('should create an untitled chat from command palette', async ({
+    page
+  }) => {
+    await page
+      .locator(
+        '#modal-command-palette li[data-command="collaborative-chat:create"]'
+      )
+      .click();
+    await fillModal(page);
+    await page.waitForCondition(
+      async () => await page.filebrowser.contents.fileExists('untitled.chat')
+    );
+    await expect(page.activity.getTabLocator('untitled.chat')).toBeVisible();
+  });
+
+  test('should not create a chat if modal is cancelled', async ({ page }) => {
+    await page
+      .locator(
+        '#modal-command-palette li[data-command="collaborative-chat:create"]'
+      )
+      .click();
+    await fillModal(page, '', 'Cancel');
+    const tab = page.getByRole('main').getByRole('tab');
+    await expect(tab).toHaveCount(1);
+  });
+
+  test('should open an existing chat', async ({ page }) => {
+    // Create a chat file
+    await page.filebrowser.contents.uploadContent('{}', 'text', FILENAME);
+
+    // open it from command palette
+    await page
+      .locator(
+        '#modal-command-palette li[data-command="collaborative-chat:open"]'
+      )
+      .click();
+    await fillModal(page, FILENAME);
+    await expect(page.activity.getTabLocator(FILENAME)).toBeVisible();
+  });
+});
+
+test.describe('#menuNew', () => {
+  test('should have an entry in main menu -> new', async ({ page }) => {
+    const menu = await page.menu.openLocator('File>New');
+    // Snapshot on list the list to avoid the menu border with transparency,
+    // which can lead to error due to background.
+    expect(await menu!.locator('> ul').screenshot()).toMatchSnapshot(
+      'menu-new.png'
+    );
+  });
+
+  test('should open modal create from the menu', async ({ page }) => {
+    await page.menu.clickMenuItem('File>New>Chat');
+    await expect(page.locator('dialog .jp-Dialog-header')).toHaveText(
+      'Create a new chat'
+    );
+  });
+});
+
+test.describe('#launcher', () => {
+  test('should have a launcher item in section', async ({ page }) => {
+    // Chat section
+    await expect(
+      page.locator('.jp-Launcher-sectionTitle:text("Chat")')
+    ).toHaveCount(1);
+
+    // Chat tile
+    const tile = page.locator('.jp-LauncherCard[data-category="Chat"]');
+    expect(await tile.screenshot()).toMatchSnapshot('launcher-tile.png');
+  });
+
+  test('should open modal create from the launcher', async ({ page }) => {
+    await page.locator('.jp-LauncherCard').getByTitle('Create a chat').click();
+    await expect(page.locator('dialog .jp-Dialog-header')).toHaveText(
+      'Create a new chat'
+    );
+  });
+});
