+++ python/jupyterlab-collaborative-chat/ui-tests/tests/commands.spec.ts-snapshots/launcher-tile-linux.png
