+++ python/jupyterlab-collaborative-chat/ui-tests/tests/commands.spec.ts-snapshots/menu-new-linux.png
