--- python/jupyterlab-collaborative-chat/ui-tests/tests/jupyterlab_collaborative_chat.spec.ts
@@ -1,376 +1,0 @@
-/*
- * Copyright (c) Jupyter Development Team.
- * Distributed under the terms of the Modified BSD License.
- */
-
-import {
-  IJupyterLabPageFixture,
-  expect,
-  galata,
-  test
-} from '@jupyterlab/galata';
-import { Contents } from '@jupyterlab/services';
-import { ReadonlyJSONObject, UUID } from '@lumino/coreutils';
-
-import { openChat, sendMessage, USER } from './test-utils';
-
-const FILENAME = 'my-chat.chat';
-const MSG_CONTENT = 'Hello World!';
-const USERNAME = USER.identity.username;
-
-test.use({
-  mockUser: USER,
-  mockSettings: { ...galata.DEFAULT_SETTINGS }
-});
-
-const fillModal = async (
-  page: IJupyterLabPageFixture,
-  text = '',
-  button: 'Ok' | 'Cancel' = 'Ok'
-): Promise<void> => {
-  const dialog = page.getByRole('dialog');
-  await dialog.getByRole('textbox').pressSequentially(text);
-  await dialog.getByRole('button').filter({ hasText: button }).click();
-};
-
-const readFileContent = async (
-  page: IJupyterLabPageFixture,
-  filename: string
-): Promise<Contents.IModel> => {
-  return await page.evaluate(async filepath => {
-    return await window.jupyterapp.serviceManager.contents.get(filepath);
-  }, filename);
-};
-
-test.describe('#commandPalette', () => {
-  const name = FILENAME.replace('.chat', '');
-
-  test.beforeEach(async ({ page }) => {
-    await page.keyboard.press('Control+Shift+c');
-  });
-
-  test.afterEach(async ({ page }) => {
-    for (let filename of ['untitled.chat', FILENAME]) {
-      if (await page.filebrowser.contents.fileExists(filename)) {
-        await page.filebrowser.contents.deleteFile(filename);
-      }
-    }
-  });
-
-  test('should have 2 commands in palette', async ({ page }) => {
-    await expect(
-      page.locator(
-        '#modal-command-palette li[data-command^="collaborative-chat"]'
-      )
-    ).toHaveCount(2);
-  });
-
-  test('should create a chat with name from command palette', async ({
-    page
-  }) => {
-    await page
-      .locator(
-        '#modal-command-palette li[data-command="collaborative-chat:create"]'
-      )
-      .click();
-    await fillModal(page, name);
-    await page.waitForCondition(
-      async () => await page.filebrowser.contents.fileExists(FILENAME)
-    );
-    await expect(page.activity.getTabLocator(FILENAME)).toBeVisible();
-  });
-
-  test('should create an untitled chat from command palette', async ({
-    page
-  }) => {
-    await page
-      .locator(
-        '#modal-command-palette li[data-command="collaborative-chat:create"]'
-      )
-      .click();
-    await fillModal(page);
-    await page.waitForCondition(
-      async () => await page.filebrowser.contents.fileExists('untitled.chat')
-    );
-    await expect(page.activity.getTabLocator('untitled.chat')).toBeVisible();
-  });
-
-  test('should not create a chat if modal is cancelled', async ({ page }) => {
-    await page
-      .locator(
-        '#modal-command-palette li[data-command="collaborative-chat:create"]'
-      )
-      .click();
-    await fillModal(page, '', 'Cancel');
-    const tab = page.getByRole('main').getByRole('tab');
-    await expect(tab).toHaveCount(1);
-  });
-
-  test('should open an existing chat', async ({ page }) => {
-    // Create a chat file
-    await page.filebrowser.contents.uploadContent('{}', 'text', FILENAME);
-
-    // open it from command palette
-    await page
-      .locator(
-        '#modal-command-palette li[data-command="collaborative-chat:open"]'
-      )
-      .click();
-    await fillModal(page, FILENAME);
-    await expect(page.activity.getTabLocator(FILENAME)).toBeVisible();
-  });
-});
-
-test.describe('#menuNew', () => {
-  test('should have an entry in main menu -> new', async ({ page }) => {
-    const menu = await page.menu.openLocator('File>New');
-    // Snapshot on list the list to avoid the menu border with transparency,
-    // which can lead to error due to background.
-    expect(await menu!.locator('> ul').screenshot()).toMatchSnapshot(
-      'menu-new.png'
-    );
-  });
-
-  test('should open modal create from the menu', async ({ page }) => {
-    await page.menu.clickMenuItem('File>New>Chat');
-    await expect(page.locator('dialog .jp-Dialog-header')).toHaveText(
-      'Create a new chat'
-    );
-  });
-});
-
-test.describe('#launcher', () => {
-  test('should have a launcher item in section', async ({ page }) => {
-    // Chat section
-    await expect(
-      page.locator('.jp-Launcher-sectionTitle:text("Chat")')
-    ).toHaveCount(1);
-
-    // Chat tile
-    const tile = page.locator('.jp-LauncherCard[data-category="Chat"]');
-    expect(await tile.screenshot()).toMatchSnapshot('launcher-tile.png');
-  });
-
-  test('should open modal create from the launcher', async ({ page }) => {
-    await page.locator('.jp-LauncherCard').getByTitle('Create a chat').click();
-    await expect(page.locator('dialog .jp-Dialog-header')).toHaveText(
-      'Create a new chat'
-    );
-  });
-});
-
-test.describe('#raw_time', () => {
-  const msg_raw_time = {
-    type: 'msg',
-    id: UUID.uuid4(),
-    sender: USERNAME,
-    body: MSG_CONTENT,
-    time: 1714116341,
-    raw_time: true
-  };
-  const msg_verif = {
-    type: 'msg',
-    id: UUID.uuid4(),
-    sender: USERNAME,
-    body: MSG_CONTENT,
-    time: 1714116341,
-    raw_time: false
-  };
-  const chatContent = {
-    messages: [msg_raw_time, msg_verif],
-    users: {}
-  };
-  chatContent.users[USERNAME] = USER.identity;
-
-  test.beforeEach(async ({ page }) => {
-    // Create a chat file with content
-    await page.filebrowser.contents.uploadContent(
-      JSON.stringify(chatContent),
-      'text',
-      FILENAME
-    );
-  });
-
-  test.afterEach(async ({ page }) => {
-    if (await page.filebrowser.contents.fileExists(FILENAME)) {
-      await page.filebrowser.contents.deleteFile(FILENAME);
-    }
-  });
-
-  test('message timestamp should be raw according to file content', async ({
-    page
-  }) => {
-    const chatPanel = await openChat(page, FILENAME);
-    const messages = chatPanel.locator(
-      '.jp-chat-messages-container .jp-chat-message'
-    );
-
-    const raw_time = messages.locator('.jp-chat-message-time').first();
-    expect(await raw_time.getAttribute('title')).toBe('Unverified time');
-    expect(await raw_time.textContent()).toMatch(/\*$/);
-
-    const verified_time = messages.locator('.jp-chat-message-time').last();
-    expect(await verified_time.getAttribute('title')).toBe('');
-    expect(await verified_time.textContent()).toMatch(/[^\*]$/);
-  });
-
-  test('time for new message should not be raw', async ({ page }) => {
-    const chatPanel = await openChat(page, FILENAME);
-    const messages = chatPanel.locator(
-      '.jp-chat-messages-container .jp-chat-message'
-    );
-
-    // Send a new message
-    await sendMessage(page, FILENAME, MSG_CONTENT);
-
-    expect(messages).toHaveCount(3);
-    await expect(
-      messages.locator('.jp-chat-message-time').last()
-    ).toHaveAttribute('title', '');
-  });
-});
-
-test.describe('#ychat', () => {
-  const chatContent = {
-    messages: [],
-    users: {}
-  };
-  chatContent.users[USERNAME] = USER.identity;
-
-  test.beforeEach(async ({ page }) => {
-    // Create a chat file with content
-    await page.filebrowser.contents.uploadContent(
-      JSON.stringify(chatContent),
-      'text',
-      FILENAME
-    );
-  });
-
-  test.afterEach(async ({ page }) => {
-    if (await page.filebrowser.contents.fileExists(FILENAME)) {
-      await page.filebrowser.contents.deleteFile(FILENAME);
-    }
-  });
-
-  test('should add an id to the chat metadata', async ({ page }) => {
-    const chatPanel = await openChat(page, FILENAME);
-    await chatPanel
-      .locator('.jp-chat-input-container')
-      .getByRole('combobox')
-      .waitFor();
-    const hasId = async () => {
-      const model = await readFileContent(page, FILENAME);
-      const content = JSON.parse(model.content) as ReadonlyJSONObject;
-      return (
-        content.metadata !== undefined &&
-        (content.metadata as ReadonlyJSONObject).id !== undefined
-      );
-    };
-    await page.waitForCondition(hasId);
-  });
-});
-
-test.describe('#outofband', () => {
-  const msg = {
-    type: 'msg',
-    id: UUID.uuid4(),
-    sender: USERNAME,
-    body: MSG_CONTENT,
-    time: 1714116341,
-    raw_time: false
-  };
-  const chatContent = {
-    messages: [msg],
-    users: {}
-  };
-  chatContent.users[USERNAME] = USER.identity;
-
-  test.beforeEach(async ({ page }) => {
-    // Create a chat file with content
-    await page.filebrowser.contents.uploadContent(
-      JSON.stringify(chatContent),
-      'text',
-      FILENAME
-    );
-  });
-
-  test.afterEach(async ({ page }) => {
-    if (await page.filebrowser.contents.fileExists(FILENAME)) {
-      await page.filebrowser.contents.deleteFile(FILENAME);
-    }
-  });
-
-  test('should update message from file', async ({ page }) => {
-    const updatedContent = 'Content updated';
-    const chatPanel = await openChat(page, FILENAME);
-    const messageContent = chatPanel
-      .locator('.jp-chat-messages-container .jp-chat-rendermime-markdown')
-      .first();
-    const newMsg = { ...msg, body: updatedContent };
-    const newContent = {
-      messages: [newMsg],
-      users: {}
-    };
-    newContent.users[USERNAME] = USER.identity;
-
-    await page.filebrowser.contents.uploadContent(
-      JSON.stringify(newContent),
-      'text',
-      FILENAME
-    );
-
-    await expect(messageContent).toHaveText(updatedContent);
-  });
-
-  test('should add a message from file', async ({ page }) => {
-    const newMsgContent = 'New message';
-    const chatPanel = await openChat(page, FILENAME);
-    const messages = chatPanel.locator(
-      '.jp-chat-messages-container .jp-chat-message'
-    );
-    const newMsg = {
-      type: 'msg',
-      id: UUID.uuid4(),
-      sender: USERNAME,
-      body: newMsgContent,
-      time: msg.time + 5,
-      raw_time: false
-    };
-    const newContent = {
-      messages: [msg, newMsg],
-      users: {}
-    };
-    newContent.users[USERNAME] = USER.identity;
-
-    await page.filebrowser.contents.uploadContent(
-      JSON.stringify(newContent),
-      'text',
-      FILENAME
-    );
-
-    await expect(messages).toHaveCount(2);
-    await expect(
-      messages.last().locator('.jp-chat-rendermime-markdown')
-    ).toHaveText(newMsgContent);
-  });
-
-  test('should delete a message from file', async ({ page }) => {
-    const chatPanel = await openChat(page, FILENAME);
-    const messageContent = chatPanel
-      .locator('.jp-chat-messages-container .jp-chat-rendermime-markdown')
-      .first();
-    const newContent = {
-      messages: [],
-      users: {}
-    };
-    newContent.users[USERNAME] = USER.identity;
-
-    await page.filebrowser.contents.uploadContent(
-      JSON.stringify(newContent),
-      'text',
-      FILENAME
-    );
-
-    await expect(messageContent).not.toBeAttached();
-  });
-});
